refactor: 계좌번호 포매팅 유틸 함수 #31
